requirements.txt
# Core Application Framework
streamlit
streamlit-webrtc

# ML & Data Processing
tensorflow
numpy
opencv-python-headless
Pillow
